wip: operations & validations refactoring
