+++ src/index.ts
@@ -3,5 +3,5 @@ export * from '@lib/errors'
 export * from '@lib/util'
 
 export { Registry as Services } from '@services/registry'
-export { Operation } from './operation'
-export { getSchema, getValidData as validate } from './validation'
+export { Operation } from '@src/operations'
+export { getSchema, getValidData as validate } from '@src/validation'
