+++ src/services/providers/aws/services/provider.ts
@@ -124,3 +124,6 @@ export const getProviderService = (): AwsProviderService =>
   )(getBaseService(PROVIDER.AWS, SERVICE_TYPE.PROVIDER))
 
 export const AwsProvider = getProviderService()
+
+export const AwsProviderDeployment = getProviderService()
+export const AwsProviderPrerequisites = getProviderService()
