+++ src/services/utils/index.ts
@@ -1,7 +1,2 @@
 export * from './getBaseService'
 export * from './getProfile'
-export * from './validation/getNameSchema'
-export * from './validation/validateEnvironment'
-export * from './validation/validateServiceLinks'
-export * from './validation/validateServiceProfile'
-export * from './validation/validateServiceProfileOverrides'
