+++ src/validation/constants.ts
@@ -1,0 +1,12 @@
+import type { Options as AjvOptions } from 'ajv'
+
+export const AJV_DEFAULTS: AjvOptions = {
+  useDefaults: true,
+  allErrors: true,
+  discriminator: false,
+  removeAdditional: true,
+  coerceTypes: true,
+  allowMatchingProperties: true,
+  strict: false,
+  $data: true,
+} as const
