+++ src/validation/utils/__tests__/validateEnvironment.test.ts
