+++ src/validation/utils/parseErrors.ts
@@ -1,0 +1,21 @@
+import { uniqBy } from 'lodash'
+import type { ValidationErrorDescriptor } from '@src/lib/errors'
+import type { ErrorObject as AjvErrorObject } from 'ajv'
+
+/**
+ * Parses Ajv errors to custom, error descriptors
+ *
+ * @param {AjvErrorObject[]} errors the raw, AJV errors available
+ * @returns {ValidationErrorDescriptor[]} the parsed errors
+ */
+export const parseErrors = (errors: AjvErrorObject[]): ValidationErrorDescriptor[] => {
+  const errs = errors
+    .filter(({ keyword }) => !['if', 'then'].includes(keyword))
+    .map(({ instancePath, message }) => {
+      const path = instancePath.replace(/\//g, '.').replace(/^\.(.*)/gi, '$1')
+      const defaultMessage = `Property ${path} is invalid`
+      return { path, message: message || defaultMessage }
+    })
+
+  return uniqBy(errs, ({ path, message }) => `${path}-${message}`)
+}
