+++ src/validation/utils/validateEnvironment.ts
